requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
openai
python-dotenv